minimize testsuite test
