+++ conan/test_package/CMakeLists.txt
@@ -48,7 +48,7 @@ target_link_libraries(${PROJECT_NAME}_mongo ${PROJECT_NAME}_objs userver::mongo)
 
 set(PYTHON_BINARY ${Python_EXECUTABLE})
 set(TESTSUITE_PYTHON_BINARY ${Python_EXECUTABLE})
-add_subdirectory(testsuite-support)
+add_subdirectory(hello_service)
 
 
 add_executable(${PROJECT_NAME}_universal test_universal.cpp)
