+++ conan/test_package/hello_service/CMakeLists.txt
@@ -1,0 +1,19 @@
+project(userver-samples-hello_service CXX)
+
+add_executable(${PROJECT_NAME} "hello_service.cpp")
+target_link_libraries(${PROJECT_NAME} userver::core)
+
+function(userver_sample_testsuite_add)
+    userver_testsuite_add(
+      SERVICE_TARGET ${PROJECT_NAME}
+      WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
+      PYTHON_BINARY ${TESTSUITE_PYTHON_BINARY}
+      PYTEST_ARGS
+      --service-config=${CMAKE_CURRENT_SOURCE_DIR}/static_config.yaml
+      --service-source-dir=${CMAKE_CURRENT_SOURCE_DIR}
+      --service-binary=${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}
+      --config-fallback=${CMAKE_CURRENT_SOURCE_DIR}/dynamic_config_fallback.json
+    )
+endfunction()
+
+userver_sample_testsuite_add()
