+++ conan/test_package/hello_service/dynamic_config_fallback.json
@@ -11,7 +11,6 @@
     "cancel-request": false,
     "update-timeout": false
   },
-  "MEANING_OF_LIFE": 42,
   "USERVER_BAGGAGE_ENABLED": false,
   "USERVER_CACHES": {},
   "USERVER_CANCEL_HANDLE_REQUEST_BY_DEADLINE": false,
