--- conan/test_package/testsuite-support/config_vars.yaml
@@ -1,1 +1,0 @@
-testsuite-enabled: false
