--- conan/test_package/testsuite-support/tests/conftest.py
@@ -1,3 +1,0 @@
-# /// [testsuite - pytest_plugins]
-pytest_plugins = ['pytest_userver.plugins.core']
-# /// [testsuite - pytest_plugins]
